Add notification Service

+++ src/main/java/fa/training/car_rental_management/controllers/PaymentController.java
@@ -157,7 +157,7 @@ public class PaymentController {
     public ResponseEntity<ApiResponse<Payment>> processSecurityDeposit(
             @RequestParam Integer bookingId,
             @RequestParam Integer payerId,
-            @RequestParam BigDecimal amount) {
+            @RequestParam Double amount) {
         try {
             log.info("Processing security deposit for booking: {}", bookingId);
